LPS-75962 Rename module
