+++ analytics-client-impl/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Analytics Client Impl
+Bundle-SymbolicName: com.liferay.analytics.client.impl
+Bundle-Version: 1.0.0
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Analytics
