--- analytics-java-client-impl/bnd.bnd
@@ -1,5 +1,0 @@
-Bundle-Name: Liferay Analytics Java Client Implementation
-Bundle-SymbolicName: com.liferay.analytics.java.client.impl
-Bundle-Version: 1.0.0
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Analytics
